# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Apvero Platform API
  version: 0.1.0
  description: Application-first control and runtime API for the Apvero self-hosted vertical slice.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/applications/{applicationId}/releases:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ApplicationId"
    get:
      tags: [Releases]
      summary: List application releases
      operationId: listReleases
      responses:
        "200":
          description: Immutable release bundles for one application.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ReleaseBundle" } }
        "404": { $ref: "#/components/responses/Problem" }
    post:
      tags: [Releases]
      summary: Create an immutable release
      operationId: createRelease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: string, pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+(?:-[a-z0-9.-]+)?$" }
                manifest: { $ref: "../schemas/release-bundle-manifest.schema.json" }
      responses:
        "201":
          description: Content-addressed release created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReleaseBundle" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/applications/{applicationId}/runs:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: [Runs]
      summary: Execute a released application
      operationId: executeRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [releaseId, input]
              properties:
                releaseId: { type: string, format: uuid }
                input: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Completed deterministic local run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunRecord" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/runs:
    get:
      tags: [Runs]
      summary: List runtime executions
      operationId: listRuns
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          description: Latest 200 workspace-scoped runs.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/RunRecord" } }
        "400": { $ref: "#/components/responses/Problem" }
...
  /api/v1/applications/{applicationId}/preview-runs:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: [Runs]
      summary: Execute an immutable preview bundle from the current application draft
      operationId: executePreviewRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [input]
              properties:
                input: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Persisted preview run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunRecord" }
        "400": { $ref: "#/components/responses/Problem" }
  /api/v1/usage:
    get:
      tags: [Runs]
      summary: Get normalized workspace usage and cost totals
      operationId: getUsage
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          description: Server-confirmed usage aggregation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsageSummary" }
...
  /api/v1/api-keys:
    get:
      tags: [Access]
      summary: List API key metadata
      operationId: listApiKeys
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          description: API key metadata without plaintext.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ApiCredential" } }
    post:
      tags: [Access]
      summary: Issue a once-visible API key
      operationId: issueApiKey
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, scopes]
              properties:
                name: { type: string, minLength: 1, maxLength: 160 }
                scopes: { type: array, minItems: 1, uniqueItems: true, items: { type: string, enum: [read, write, admin] } }
                expiresAt: { type: [string, "null"], format: date-time }
      responses:
        "201":
          description: API key plaintext is returned exactly once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedApiCredential" }
        "400": { $ref: "#/components/responses/Problem" }
  /api/v1/api-keys/{credentialId}:
    delete:
      tags: [Access]
      summary: Revoke an API key
      operationId: revokeApiKey
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: credentialId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: API key revoked. }
        "400": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Apvero API Key
  parameters:
    WorkspaceId:
      name: X-Apvero-Workspace-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Explicit workspace security and data scope.
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Problem:
      description: RFC 9457 problem detail with a stable Apvero code.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ReleaseBundle:
      type: object
      required: [id, tenantId, workspaceId, applicationId, version, artifactDigest, manifest, status, purpose, createdAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        version: { type: string }
        artifactDigest: { type: string, pattern: "^[a-f0-9]{64}$" }
        manifest: { $ref: "../schemas/release-bundle-manifest.schema.json" }
        status: { type: string, enum: [RELEASED, RETIRED] }
        purpose: { type: string, enum: [PRODUCTION, PREVIEW] }
        expiresAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    RunRecord:
      type: object
      required: [id, tenantId, workspaceId, applicationId, releaseBundleId, status, providerId, input, output, latencyMs, promptTokens, completionTokens, costMicros, traceId, createdAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        releaseBundleId: { type: string, format: uuid }
        status: { type: string, enum: [SUCCEEDED, FAILED] }
        providerId: { type: string }
        input: { type: object, additionalProperties: true }
        output: { type: object, additionalProperties: true }
        latencyMs: { type: integer, format: int64, minimum: 0 }
        promptTokens: { type: integer, minimum: 0 }
        completionTokens: { type: integer, minimum: 0 }
        costMicros: { type: integer, format: int64, minimum: 0 }
        traceId: { type: string }
        failureCategory: { type: [string, "null"] }
        failureMessage: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    UsageSummary:
      type: object
      additionalProperties: false
      required: [runs, successfulRuns, failedRuns, promptTokens, completionTokens, totalTokens, costMicros, averageLatencyMs]
      properties:
        runs: { type: integer, format: int64, minimum: 0 }
        successfulRuns: { type: integer, format: int64, minimum: 0 }
        failedRuns: { type: integer, format: int64, minimum: 0 }
        promptTokens: { type: integer, format: int64, minimum: 0 }
        completionTokens: { type: integer, format: int64, minimum: 0 }
        totalTokens: { type: integer, format: int64, minimum: 0 }
        costMicros: { type: integer, format: int64, minimum: 0 }
        averageLatencyMs: { type: number, minimum: 0 }
...
    ApiCredential:
      type: object
      additionalProperties: false
      required: [id, tenantId, workspaceId, name, prefix, scopes, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid }
        name: { type: string }
        prefix: { type: string }
        scopes: { type: array, items: { type: string, enum: [read, write, admin] } }
        status: { type: string, enum: [ACTIVE, REVOKED] }
        expiresAt: { type: [string, "null"], format: date-time }
        lastUsedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    IssuedApiCredential:
      type: object
      additionalProperties: false
      required: [credential, plaintext]
      properties:
        credential: { $ref: "#/components/schemas/ApiCredential" }
        plaintext: { type: string, writeOnly: true }
    Problem:
      type: object
      required: [type, title, status, detail, code, timestamp]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        timestamp: { type: string, format: date-time }
        fields: { type: object, additionalProperties: { type: string } }